Refactor compute copy helper header per convention
