+++ examples/cuda/bubblechart.cu
@@ -12,7 +12,7 @@
 #include <curand_kernel.h>
 #include <forge.h>
 #define USE_FORGE_CUDA_COPY_HELPERS
-#include <ComputeCopy.h>
+#include <fg/compute_copy.h>
 #include <cstdio>
 #include <iostream>
