+++ examples/cuda/plot3.cu
@@ -11,7 +11,7 @@
 #include <cuda_runtime.h>
 #include <forge.h>
 #define USE_FORGE_CUDA_COPY_HELPERS
-#include <ComputeCopy.h>
+#include <fg/compute_copy.h>
 #include <cstdio>
 #include <iostream>
